+++ src/pages/Dev/components/BottomBar/index.tsx
@@ -1,5 +1,3 @@
-import jsPDF from 'jspdf'
-import type { RGBAData } from 'jspdf'
 import Icon from '@ant-design/icons'
 import cssSVG from '@/assets/svg/dev/css.svg?react'
 import centerSVG from '@/assets/svg/dev/center.svg?react'
@@ -10,7 +8,6 @@ import normalSVG from '@/assets/svg/dev/normal.svg?react'
 import codeSVG from '@/assets/svg/dev/code.svg?react'
 import styles from './index.module.scss'
 import { ConfigProvider, Switch, Tooltip } from 'antd'
-import { sleep } from '@/utils'
 
 // type toCanvas = {
 //   base64URL: string
@@ -25,15 +22,7 @@ type barType = {
   resetWheel: () => void
   isDev: boolean
   setIsDev: (val: boolean) => void
-  dom2Canvas: () => Promise<
-    | {
-        base64URL: string
-        width: number
-        height: number
-      }
-    | undefined
-  >
-  setWheel: (wheel: number) => void
+  savePDF: () => Promise<any>
 }
 
 const BottomBar = ({
@@ -43,48 +32,47 @@ const BottomBar = ({
   resetWheel,
   isDev,
   setIsDev,
-  dom2Canvas,
-  setWheel,
+  savePDF,
 }: barType) => {
-  const addImage = (
-    _x: number,
-    _y: number,
-    pdfInstance: jsPDF,
-    base_data:
-      | string
-      | HTMLImageElement
-      | HTMLCanvasElement
-      | Uint8Array
-      | RGBAData,
-    _width: number,
-    _height: number
-  ) => {
-    pdfInstance.addImage(base_data, 'JPEG', _x, _y, _width, _height)
-  }
+  // const addImage = (
+  //   _x: number,
+  //   _y: number,
+  //   pdfInstance: jsPDF,
+  //   base_data:
+  //     | string
+  //     | HTMLImageElement
+  //     | HTMLCanvasElement
+  //     | Uint8Array
+  //     | RGBAData,
+  //   _width: number,
+  //   _height: number
+  // ) => {
+  //   pdfInstance.addImage(base_data, 'JPEG', _x, _y, _width, _height)
+  // }
 
-  const savePDF = async () => {
-    setWheel(1)
-    await sleep()
-    const {
-      base64URL,
-      height: canvasHeight,
-      width: canvasWidth,
-    } = (await dom2Canvas()) || {}
-    console.log(base64URL, canvasHeight, canvasWidth)
-    const pdfInstance = new jsPDF({
-      orientation: 'portrait',
-      unit: 'px',
-      format: 'a4',
-    })
-    if (!base64URL || !canvasHeight || !canvasWidth) return
-    const pageWidth = pdfInstance.internal.pageSize.getWidth()
+  // const savePDF = async () => {
+  //   setWheel(1)
+  //   await sleep()
+  //   const {
+  //     base64URL,
+  //     height: canvasHeight,
+  //     width: canvasWidth,
+  //   } = (await dom2Canvas()) || {}
+  //   console.log(base64URL, canvasHeight, canvasWidth)
+  //   const pdfInstance = new jsPDF({
+  //     orientation: 'portrait',
+  //     unit: 'px',
+  //     format: 'a4',
+  //   })
+  //   if (!base64URL || !canvasHeight || !canvasWidth) return
+  //   const pageWidth = pdfInstance.internal.pageSize.getWidth()
 
-    // 计算等比缩放后的尺寸
-    const scale = pageWidth / canvasWidth
-    const imgHeight = canvasHeight * scale
-    addImage(0, 0, pdfInstance, base64URL, pageWidth, imgHeight)
-    pdfInstance.save(`resume-${new Date().getTime()}`)
-  }
+  //   // 计算等比缩放后的尺寸
+  //   const scale = pageWidth / canvasWidth
+  //   const imgHeight = canvasHeight * scale
+  //   addImage(0, 0, pdfInstance, base64URL, pageWidth, imgHeight)
+  //   pdfInstance.save(`resume-${new Date().getTime()}`)
+  // }
 
   const iconArr = [
     {
